Added VocabularyTestInstrumentedTest
